Extract db iterator class.
